Добавляет правки в код, и разбивает на модули
