+++ js/data.js
@@ -1,0 +1,36 @@
+//Исходные данные
+const NAMES = [
+  'Елена',
+  'Екатерина',
+  'Ольга',
+  'Дмитрий',
+  'Николай',
+  'Андрей',
+  'Иван',
+  'Александр',
+  'Анна',
+  'Татьяна',
+  'Михаил'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const DESCRIPTIONS = [
+  'Интересная фотография.',
+  'На этой фотографии изображено множество деталей.',
+  'Мне кажется, фотографию сделал профессионал',
+  'На фотографии мы видим весёлых котиков',
+  'Перед нами чудесный летний лес, на который хочется смотреть вечно.',
+  'Они выглядят счастливыми.',
+  'Я считаю, что снимок получился удачным.',
+  'Мне понравилась фотография, потому что она передает эмоции и атмосферу.'
+];
+
+export {NAMES, MESSAGES, DESCRIPTIONS};
